Split tracing internal plumbing modules
